% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{clustering_statistics}
\alias{clustering_statistics}
\title{clustering stats}
\usage{
clustering_statistics(
  input_data = data[[2]],
  k_start = 2,
  k_end = 5,
  type = "kmeans"
)
}
\value{
stats
}
\description{
clustering stats
}
\examples{
eg = clustering_statistics();
}
