% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{read_data_clinical}
\alias{read_data_clinical}
\title{read clinical data}
\usage{
read_data_clinical(file = NULL)
}
\value{
read
}
\description{
read clinical data
}
\examples{
eg = read_data_clinical();
}
