% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_plots.R
\name{render_detailed_silhouette}
\alias{render_detailed_silhouette}
\title{render detailed silhouette}
\usage{
render_detailed_silhouette(
  input_data = data[[2]],
  k = 2,
  colors = wes_colors,
  type = c("kmeans", "dendrogram")
)
}
\value{
render
}
\description{
render detailed silhouette
}
\examples{
eg = render_detailed_silhouette();
}
