% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_plots.R
\name{render_summary_data}
\alias{render_summary_data}
\title{render summary data}
\usage{
render_summary_data(
  clinical = clinical,
  variable = colnames(clinical),
  colors = wes_colors
)
}
\value{
render
}
\description{
render summary data
}
\examples{
eg = render_summary_data();
}
