% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_plots.R
\name{renderpca}
\alias{renderpca}
\title{render pca}
\usage{
renderpca(
  input_data = data[[2]],
  pcX = 1,
  pcY = 2,
  k = 2,
  colors = wes_colors,
  type = "kmeans"
)
}
\value{
render
}
\description{
render pca
}
\examples{
eg = renderpca();
}
