% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_plots.R
\name{rendersilhouette}
\alias{rendersilhouette}
\title{render  silhouette}
\usage{
rendersilhouette(
  multi_k = clustering_statistics(data[[2]]),
  colors = wes_colors
)
}
\value{
render
}
\description{
render  silhouette
}
\examples{
eg = rendersilhouette();
}
