% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_plots.R
\name{renderstwss}
\alias{renderstwss}
\title{render  twss}
\usage{
renderstwss(multi_k = clustering_statistics(data[[2]]), colors = wes_colors)
}
\value{
render
}
\description{
render  twss
}
\examples{
eg = renderstwss();
}
