% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_plots.R
\name{switch_clustering_plots}
\alias{switch_clustering_plots}
\title{switch clustering plots}
\usage{
switch_clustering_plots(
  multi_k = multi_k,
  data = data,
  clustering_metrics = "",
  k = 2
)
}
\value{
plots
}
\description{
switch clustering plots
}
\examples{
eg = switch_clustering_plots();
}
